+++ src/app/mypage/edit/page.tsx
@@ -1,4 +1,4 @@
-import EditForm from '@components/mypage/MyEditForm';
+import EditForm from '@/components/mypage/edit';
 
 export default function mypageEdit() {
   return (
